refactor(opentrons): move Flex driver into opentrons.flex subpackage
Mirror Rick's opentrons/ot2/ layout: relocate the Flex driver + support
modules (flex, flex_head, flex_gripper, flex_wire, labware_definitions,
checks, envelope, pipette_defaults) and their tests into a new
pylabrobot/opentrons/flex/ subpackage, re-pointing intra-package imports.

Pure move, no behavior change: the flat robot.py/transport.py wire layer
stays in the parent for now (replaced by the api/run rewrite next). Flex
tests (Chatterbox-based) unchanged and green.
