+++ pylabrobot/opentrons/flex/__init__.py
@@ -1,0 +1,11 @@
+from pylabrobot.opentrons.flex.flex import OpentronsFlex
+from pylabrobot.opentrons.flex.flex_gripper import FlexGripper
+from pylabrobot.opentrons.flex.flex_head import FlexHead1, FlexHead8, FlexHead96
+
+__all__ = [
+  "FlexGripper",
+  "FlexHead1",
+  "FlexHead8",
+  "FlexHead96",
+  "OpentronsFlex",
+]
