+++ pylabrobot/opentrons/flex/checks.py
@@ -8,12 +8,12 @@ dispatch, and PLR owns the check rather than delegating it to the vendor.
 
 from typing import List, Optional
 
-from pylabrobot.opentrons.envelope import (
+from pylabrobot.opentrons.flex.envelope import (
   FLEX_8CH_NOZZLE_A1_Y,
   FLEX_ENVELOPE,
   nozzle_offset_y,
 )
-from pylabrobot.opentrons.envelope import (
+from pylabrobot.opentrons.flex.envelope import (
   FLEX_8CH_NOZZLE_H1_Y as FLEX_8CH_NOZZLE_H1_Y,  # re-exported for callers importing from checks
 )
 from pylabrobot.opentrons.robot import PipetteInfo
