+++ pylabrobot/opentrons/flex/flex_envelope_tests.py
@@ -10,10 +10,10 @@ import asyncio
 import unittest
 from typing import List, Tuple
 
-from pylabrobot.opentrons.checks import traversal_z
-from pylabrobot.opentrons.envelope import FLEX_ENVELOPE
-from pylabrobot.opentrons.flex import OpentronsFlex
-from pylabrobot.opentrons.flex_head import FlexHead8
+from pylabrobot.opentrons.flex.checks import traversal_z
+from pylabrobot.opentrons.flex.envelope import FLEX_ENVELOPE
+from pylabrobot.opentrons.flex.flex import OpentronsFlex
+from pylabrobot.opentrons.flex.flex_head import FlexHead8
 from pylabrobot.opentrons.transport import ChatterboxTransport
 from pylabrobot.resources import cor_96_wellplate_360uL_Fb
 from pylabrobot.resources.opentrons.flex_deck import FlexDeck
